refactor service to move each req to separate file
