+++ service.go
@@ -1,28 +1,26 @@
 package relayproxy
 
 import (
-	"bytes"
 	"context"
 	"encoding/json"
 	"errors"
 	"fmt"
 	"io"
 	"math/big"
-	"math/rand"
 	"net"
-	"net/http"
 	"strconv"
 	"strings"
 	"sync"
-	"sync/atomic"
 	"time"
 
 	eth2Api "github.com/attestantio/go-eth2-client/api"
 	"github.com/attestantio/go-eth2-client/spec/bellatrix"
 	"github.com/attestantio/go-eth2-client/spec/phase0"
+	relaygrpc "github.com/bloXroute-Labs/relay-grpc"
+	"github.com/bloXroute-Labs/relayproxy/common"
+	"github.com/bloXroute-Labs/relayproxy/fluentstats"
 	gethcommon "github.com/ethereum/go-ethereum/common"
 	"github.com/flashbots/go-boost-utils/bls"
-	"github.com/flashbots/go-boost-utils/ssz"
 	"github.com/google/uuid"
 	"github.com/patrickmn/go-cache"
 	"github.com/rs/zerolog"
@@ -32,13 +30,6 @@ import (
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/metadata"
 	"google.golang.org/grpc/status"
-	"google.golang.org/protobuf/types/known/timestamppb"
-
-	relaygrpc "github.com/bloXroute-Labs/relay-grpc"
-	"github.com/bloXroute-Labs/relayproxy/common"
-	"github.com/bloXroute-Labs/relayproxy/fastjson"
-	"github.com/bloXroute-Labs/relayproxy/fluentstats"
-	"github.com/bloXroute-Labs/relayproxy/httpclient"
 )
 
 const (
@@ -162,142 +153,6 @@ func NewService(opts ...ServiceOption) *Service {
 	return svc
 }
 
-func (s *Service) RegisterValidator(ctx context.Context, log *zerolog.Logger, outgoingCtx context.Context, in *RegistrationParams) (any, error) {
-	var (
-		errChan  = make(chan *ErrorResp, len(s.clients))
-		respChan = make(chan *relaygrpc.RegisterValidatorResponse, len(s.clients))
-		_err     *ErrorResp
-	)
-	timer := time.NewTimer(regRequestTimeout)
-	defer timer.Stop()
-
-	parentSpan := trace.SpanFromContext(ctx)
-	ctx = trace.ContextWithSpan(outgoingCtx, parentSpan)
-	ctx = metadata.AppendToOutgoingContext(ctx, "authorization", in.AuthHeader)
-	ctx, span := s.tracer.Start(ctx, "registerValidator-start")
-	defer span.End()
-
-	id := uuid.NewString()
-
-	*log = log.With().
-		Str("method", "registerValidator").
-		Str("clientIP", in.ClientIP).
-		Str("reqID", id).
-		Str("traceID", parentSpan.SpanContext().TraceID().String()).
-		Str("in.ValidatorID", in.ValidatorID).
-		Str("accountID", in.AccountID).
-		Str("authHeader", in.AuthHeader).
-		Bool("proposerMevProtect", in.ProposerMevProtect).
-		Time("receivedAt", in.ReceivedAt).
-		Logger()
-	log.Info().Msg("received registration")
-	parentSpan.SetAttributes(
-		attribute.String("method", "registerValidator"),
-		attribute.String("clientIP", in.ClientIP),
-		attribute.String("reqID", id),
-		attribute.String("in.ValidatorID", in.ValidatorID),
-		attribute.String("traceID", parentSpan.SpanContext().TraceID().String()),
-		attribute.Int64("receivedAt", in.ReceivedAt.Unix()),
-		attribute.String("authHeader", in.AuthHeader),
-		attribute.Bool("proposerMevProtect", in.ProposerMevProtect),
-	)
-
-	req := &relaygrpc.RegisterValidatorRequest{
-		ReqId:              id,
-		Payload:            in.Payload,
-		ClientIp:           in.ClientIP,
-		Version:            s.version,
-		ReceivedAt:         timestamppb.New(in.ReceivedAt),
-		AuthHeader:         in.AuthHeader,
-		SecretToken:        s.secretToken,
-		ComplianceList:     in.ComplianceList,
-		ProposerMevProtect: in.ProposerMevProtect,
-		SkipOptimism:       in.SkipOptimism,
-	}
-
-	ctx, spanWait := s.tracer.Start(ctx, "RegisterValidator-waitForResponse")
-	go func(_ctx context.Context, req *relaygrpc.RegisterValidatorRequest) {
-		out, err := s.registerValidatorForClient(ctx, req)
-		if err != nil {
-			errChan <- err
-			return
-		}
-		respChan <- out
-	}(ctx, req)
-	spanWait.End(trace.WithTimestamp(time.Now()))
-
-	ctx, spanSuccess := s.tracer.Start(ctx, "RegisterValidator-waitForSuccessfulResponse")
-	select {
-	case <-ctx.Done():
-		return nil, toErrorResp(http.StatusInternalServerError, ctx.Err().Error())
-	case _err = <-errChan:
-		// first error captured
-	case <-respChan:
-		return struct{}{}, nil
-	case <-timer.C:
-		log.Error().Msg("timer hit: relay request timeout")
-		return struct{}{}, nil
-	}
-	spanSuccess.End(trace.WithTimestamp(time.Now()))
-
-	if _err != nil {
-		if _err.Code == http.StatusRequestTimeout {
-			log.Info().Msg("relay request timeout")
-			return struct{}{}, nil
-		}
-	}
-
-	return nil, _err
-}
-
-func (s *Service) registerValidatorForClient(_ctx context.Context, req *relaygrpc.RegisterValidatorRequest) (*relaygrpc.RegisterValidatorResponse, *ErrorResp) {
-	_ctx, regSpan := s.tracer.Start(_ctx, "registerValidator-registerValidatorForClient")
-	var (
-		out *relaygrpc.RegisterValidatorResponse
-		err error
-	)
-
-	for range s.registrationClients {
-		req.NodeId = s.nodeID
-
-		s.registrationRelayMutex.Lock()
-		selectedRelay := s.registrationClients[s.currentRegistrationRelayIndex]
-		s.currentRegistrationRelayIndex = (s.currentRegistrationRelayIndex + 1) % len(s.registrationClients)
-		s.registrationRelayMutex.Unlock()
-
-		out, err = selectedRelay.SafeClient.RegisterValidator(_ctx, req)
-		url := selectedRelay.SafeClient.URL
-
-		if err != nil || out == nil || out.Code != uint32(codes.OK) {
-			s.logger.Warn().Str("url", url).Err(err).Msg("failed to register validator")
-			continue
-		}
-
-		regSpan.SetStatus(otelcodes.Ok, "relay returned success response code")
-		regSpan.End()
-		return out, nil
-	}
-
-	regSpan.SetStatus(otelcodes.Error, "relay registration failed")
-	regSpan.End()
-
-	if err != nil {
-		if strings.Contains(err.Error(), errContextDeadlineString) {
-			return nil, toErrorResp(http.StatusRequestTimeout, err.Error())
-		}
-		return nil, toErrorResp(http.StatusInternalServerError, err.Error())
-	}
-
-	if out == nil {
-		return nil, toErrorResp(http.StatusInternalServerError, "empty response from relay")
-	}
-
-	if out.Code != uint32(codes.OK) {
-		return nil, toErrorResp(http.StatusBadRequest, "relay returned failure response code"+strconv.FormatUint(uint64(out.Code), 10))
-	}
-
-	return nil, toErrorResp(http.StatusInternalServerError, "no relay client available")
-}
 func (s *Service) StartStreamHeaders(ctx context.Context, wg *sync.WaitGroup) {
 
 	for _, client := range s.streamingClients {
@@ -630,1126 +485,6 @@ func (s *Service) StreamHeader(ctx context.Context, client *common.Client, paren
 
 	s.logger.Warn().Fields(logMetric.GetFields()).Msg("closing connection")
 	return nil, nil
-}
-
-func (s *Service) GetHeader(ctx context.Context, log *zerolog.Logger, in *HeaderRequestParams) (json.RawMessage, *common.OnHeaderDeliveredParams, error) {
-	id := uuid.NewString()
-	parentSpan := trace.SpanFromContext(ctx)
-	ctx = trace.ContextWithSpan(context.Background(), parentSpan)
-	ctx, span := s.tracer.Start(ctx, "getHeader-start")
-	defer span.End()
-
-	k := "slot-" + in.Slot + "-parentHash-" + in.ParentHash
-
-	delayCtx, delayGetHeaderSpan := s.tracer.Start(ctx, "getHeader-delayGetHeader")
-
-	delayGetHeaderResponse, err := s.DelayGetHeader(delayCtx, DelayGetHeaderParams{
-		ReceivedAt:          in.ReceivedAt,
-		Slot:                in.Slot,
-		AccountID:           in.AccountID,
-		Cluster:             in.Cluster,
-		UserAgent:           in.UserAgent,
-		ClientIP:            in.ClientIP,
-		SlotWithParentHash:  k,
-		BoostSendTimeUnixMS: in.GetHeaderStartTimeUnixMS,
-		Latency:             in.Latency,
-	})
-	delayGetHeaderSpan.End(trace.WithTimestamp(time.Now()))
-
-	_, preStoringHeaderSpan := s.tracer.Start(ctx, "getHeader-preStoringHeaderSpan")
-	sleep := delayGetHeaderResponse.Sleep // TODO: refactor for the error handling
-	maxSleep := delayGetHeaderResponse.MaxSleep
-	slotStartTime := delayGetHeaderResponse.SlotStartTime
-	latency := delayGetHeaderResponse.Latency
-
-	startTime := time.Now().UTC()
-
-	*log = log.With().
-		Str("method", getHeader).
-		Str("reqID", id).
-		Str("key", k).
-		Str("slot", in.Slot).
-		Int64("slotStartTimeUnix", slotStartTime.Unix()).
-		Str("slotStartTime", slotStartTime.UTC().String()).
-		Int64("sleep", sleep).
-		Int64("maxSleep", maxSleep).
-		Logger()
-
-	parentSpan.SetAttributes(
-		attribute.String("method", getHeader),
-		attribute.String("req", id),
-		attribute.Int64("receivedAt", in.ReceivedAt.Unix()),
-		attribute.String("key", k),
-		attribute.String("slot", in.Slot),
-		attribute.Int64("slotStartTimeUnix", slotStartTime.Unix()),
-		attribute.String("slotStartTime", slotStartTime.UTC().String()),
-		attribute.Int64("sleep", sleep),
-		attribute.Int64("maxSleep", maxSleep),
-	)
-
-	log.Info().Msg("received getHeader")
-	if err != nil {
-		preStoringHeaderSpan.End(trace.WithTimestamp(time.Now()))
-		return nil, nil, toErrorResp(http.StatusNoContent, err.Error())
-	}
-
-	_, parseUintHeaderSpan := s.tracer.Start(ctx, "getHeader-parseUint")
-	_slot, err := fastParseUint(in.Slot)
-	if err != nil {
-		parseUintHeaderSpan.End(trace.WithTimestamp(time.Now()))
-		preStoringHeaderSpan.End(trace.WithTimestamp(time.Now()))
-		return nil, nil, toErrorResp(http.StatusNoContent, errInvalidSlot.Error())
-	}
-
-	parseUintHeaderSpan.End(trace.WithTimestamp(time.Now()))
-
-	_, slotTimeMeasureSpan := s.tracer.Start(ctx, "getHeader-slotTimeMeasure")
-	msIntoSlotIncludingDelay := time.Since(slotStartTime).Milliseconds()
-	msIntoSlot := in.ReceivedAt.Sub(slotStartTime).Milliseconds() // without sleep and using received at
-	*log = log.With().
-		Int64("msIntoSlot", msIntoSlot).
-		Int64("msIntoSlotIncludingDelay", msIntoSlotIncludingDelay).
-		Logger()
-
-	parentSpan.SetAttributes(
-		attribute.Int64("msTntoSlot", msIntoSlot),
-		attribute.Int64("msIntoSlotIncludingDelay", msIntoSlotIncludingDelay),
-	)
-
-	slotTimeMeasureSpan.End(trace.WithTimestamp(time.Now()))
-
-	preStoringHeaderSpan.End(trace.WithTimestamp(time.Now()))
-
-	_, storingHeaderSpan := s.tracer.Start(ctx, "getHeader-storingHeader")
-	//TODO: send fluentd stats for StatusNoContent error cases
-
-	if len(in.PubKey) != 98 {
-		storingHeaderSpan.End(trace.WithTimestamp(time.Now()))
-		return nil, nil, toErrorResp(http.StatusNoContent, errInvalidPubkey.Error())
-	}
-
-	if len(in.ParentHash) != 66 {
-		storingHeaderSpan.End(trace.WithTimestamp(time.Now()))
-		return nil, nil, toErrorResp(http.StatusNoContent, errInvalidHash.Error())
-	}
-
-	fetchGetHeaderStartTime := time.Now().UTC()
-	keyForCachingBids := s.keyForCachingBids(_slot, in.ParentHash, in.PubKey)
-	slotBestHeader, err := s.GetTopBuilderBid(keyForCachingBids)
-	fetchGetHeaderDurationMS := time.Since(fetchGetHeaderStartTime).Milliseconds()
-	headerReqDuration := time.Since(in.ReceivedAt)
-	statsUserAgent := in.UserAgent
-	if in.Cluster != "" {
-		statsUserAgent += "/" + in.Cluster
-	}
-
-	if slotBestHeader == nil || err != nil {
-		msg := fmt.Sprintf("header value is not present for the requested key %v", keyForCachingBids)
-		span.AddEvent("Header value is not present", trace.WithAttributes(attribute.String("msg", msg)))
-		go func() {
-			headerStats := GetHeaderStatsRecord{
-				RequestReceivedAt:        in.ReceivedAt,
-				FetchGetHeaderStartTime:  fetchGetHeaderStartTime.String(),
-				FetchGetHeaderDurationMS: fetchGetHeaderDurationMS,
-				Duration:                 time.Since(startTime),
-				MsIntoSlot:               msIntoSlot,
-				ParentHash:               in.ParentHash,
-				PubKey:                   in.PubKey,
-				BlockHash:                "",
-				ReqID:                    id,
-				ClientIP:                 in.ClientIP,
-				BlockValue:               "",
-				Succeeded:                false,
-				NodeID:                   s.nodeID,
-				Slot:                     int64(_slot),
-				AccountID:                in.AccountID,
-				ValidatorID:              in.ValidatorID,
-				Latency:                  latency,
-				UserAgent:                statsUserAgent,
-			}
-			s.fluentD.LogToFluentD(fluentstats.Record{
-				Type: TypeRelayProxyGetHeader,
-				Data: headerStats,
-			}, time.Now().UTC(), s.nodeID, StatsRelayProxyGetHeader)
-		}()
-		return nil, nil, toErrorResp(http.StatusNoContent, "Header value is not present")
-	}
-	if slotBestHeader.AccountID != "" {
-		in.AccountID = slotBestHeader.AccountID
-		if s.accountsLists.AccountIDToInfo[in.AccountID] != nil &&
-			s.accountsLists.AccountIDToInfo[in.AccountID].UseAccountAsValidator {
-			in.ValidatorID = in.AccountID
-		}
-	}
-	uKey := fmt.Sprintf("slot_%v_bHash_%v_pHash_%v", in.Slot, slotBestHeader.BlockHash, in.ParentHash) // TODO:add pubkey
-	blockValue := new(big.Int).SetBytes(slotBestHeader.Value)
-	*log = log.With().
-		Str("blockHash", slotBestHeader.BlockHash).
-		Str("blockValue", blockValue.String()).
-		Str("uniqueKey", uKey).
-		Logger()
-	parentSpan.SetAttributes(
-		attribute.String("blockHash", slotBestHeader.BlockHash),
-		attribute.String("blockValue", blockValue.String()),
-		attribute.String("uniqueKey", uKey),
-	)
-	storingHeaderSpan.End(trace.WithTimestamp(time.Now()))
-
-	go func() {
-		slotStats := SlotStatsRecord{
-			HeaderReqID:               id,
-			HeaderReqReceivedAt:       in.ReceivedAt,
-			HeaderReqDuration:         headerReqDuration, // this is not used
-			HeaderReqDurationInMs:     headerReqDuration.Milliseconds(),
-			HeaderDelayInMs:           sleep,
-			HeaderMaxDelayInMs:        maxSleep,
-			HeaderMsIntoSlot:          msIntoSlot,
-			HeaderMsIntoSlotWithDelay: msIntoSlotIncludingDelay,
-			HeaderSucceeded:           true,
-			HeaderDeliveredBlockHash:  slotBestHeader.BlockHash,
-			HeaderBlockValue:          weiToEther(blockValue),
-			HeaderUserAgent:           statsUserAgent,
-			HeaderStartTimeUnixMs:     in.GetHeaderStartTimeUnixMS,
-			Slot:                      _slot,
-			SlotStartTime:             slotStartTime,
-			ParentHash:                in.ParentHash,
-			PubKey:                    in.PubKey,
-			ClientIP:                  in.ClientIP,
-			NodeID:                    s.nodeID,
-			AccountID:                 in.AccountID,
-			ValidatorID:               in.ValidatorID,
-			GetHeaderLatency:          latency,
-			HeaderSlotUID:             in.SlotUID,
-		}
-
-		if v, ok := s.slotStats.Get(k); !ok {
-			slotStatsSlice := make([]SlotStatsRecord, 0, 5)
-			slotStatsSlice = append(slotStatsSlice, slotStats)
-			s.slotStats.Set(k, slotStatsSlice, cache.DefaultExpiration)
-
-			s.slotStatsEventCh <- slotStatsEvent{
-				Slot:      int64(_slot),
-				SlotKey:   k,
-				UserAgent: in.UserAgent,
-			}
-
-		} else {
-			slotStatsSlice := v.([]SlotStatsRecord)
-			slotStatsSlice = append(slotStatsSlice, slotStats)
-			s.slotStats.Set(k, slotStatsSlice, cache.DefaultExpiration)
-		}
-
-		headerStats := GetHeaderStatsRecord{
-			RequestReceivedAt:        in.ReceivedAt,
-			FetchGetHeaderStartTime:  fetchGetHeaderStartTime.String(),
-			FetchGetHeaderDurationMS: fetchGetHeaderDurationMS,
-			Duration:                 time.Since(in.ReceivedAt),
-			MsIntoSlot:               msIntoSlot,
-			ParentHash:               in.ParentHash,
-			PubKey:                   in.PubKey,
-			BlockHash:                slotBestHeader.BlockHash,
-			ReqID:                    id,
-			ClientIP:                 in.ClientIP,
-			BlockValue:               weiToEther(blockValue),
-			Succeeded:                true,
-			NodeID:                   s.nodeID,
-			Slot:                     int64(_slot),
-			AccountID:                in.AccountID,
-			ValidatorID:              in.ValidatorID,
-			Latency:                  latency,
-			UserAgent:                statsUserAgent,
-			SlotUID:                  in.SlotUID,
-			HeaderStartTimeUnixMs:    in.GetHeaderStartTimeUnixMS,
-		}
-		s.fluentD.LogToFluentD(fluentstats.Record{
-			Type: TypeRelayProxyGetHeader,
-			Data: headerStats,
-		}, time.Now().UTC(), s.nodeID, StatsRelayProxyGetHeader)
-	}()
-
-	// send in payload to pre fetcher event
-	s.preFetchPayloadChan <- preFetcherFields{
-		clientIP:        in.ClientIP,
-		authHeader:      in.AuthHeader,
-		slot:            _slot,
-		parentHash:      in.ParentHash,
-		blockHash:       slotBestHeader.BlockHash,
-		proposerPubKey:  in.PubKey,
-		builderPubKey:   slotBestHeader.BuilderPubkey,
-		blockValue:      weiToEther(blockValue),
-		client:          slotBestHeader.Client,
-		payloadFetchUrl: slotBestHeader.PayloadFetchUrl,
-	}
-
-	signedHeaderResponse, prevSigned, err := slotBestHeader.GetSignedHeaderResponse(s.secretKey, &s.publicKey, s.builderSigningDomain)
-	if err != nil {
-		log.Error().Err(err).Msg("failed to get signed header")
-	}
-	if prevSigned {
-		log.Debug().Msg("previously signed header")
-	} else {
-		log.Debug().Msg("newly signed header")
-	}
-
-	onHeaderDeliveredParams := &common.OnHeaderDeliveredParams{
-		SignedHeaderResponse:     signedHeaderResponse,
-		Slot:                     _slot,
-		GetHeaderRequestID:       in.SlotUID,
-		ProposerPubkey:           in.PubKey,
-		GetHeaderStartTimeUnixMS: in.GetHeaderStartTimeUnixMS,
-		ExtraData:                slotBestHeader.BuilderExtraData,
-	}
-
-	return json.RawMessage(signedHeaderResponse), onHeaderDeliveredParams, nil
-}
-
-func (s *Service) StartPreFetcher(ctx context.Context) {
-	for fields := range s.preFetchPayloadChan {
-		go func(fields preFetcherFields) {
-			_ctx, cancel := context.WithTimeout(ctx, preFetcherRequestTimeout)
-			defer cancel()
-			s.PreFetchGetPayload(_ctx, fields)
-		}(fields)
-	}
-}
-
-func (s *Service) PreFetchGetPayload(ctx context.Context, fields preFetcherFields) {
-	var clientURL string
-	startTime := time.Now().UTC()
-	id := uuid.NewString()
-	parentSpan := trace.SpanFromContext(ctx)
-	ctx = trace.ContextWithSpan(context.Background(), parentSpan)
-	ctx = metadata.AppendToOutgoingContext(ctx, "authorization", s.authKey)
-	spanctx, span := s.tracer.Start(ctx, "preFetchGetPayload-start")
-	uKey := fmt.Sprintf("slot_%v_bHash_%v_pHash_%v", fields.slot, fields.blockHash, fields.parentHash)
-	defer func() {
-		span.SetAttributes(
-			attribute.String("method", preFetchPayload),
-			attribute.String("clientIP", fields.clientIP),
-			attribute.String("clientURL", clientURL),
-			attribute.String("reqID", id),
-			attribute.Int64("receivedAt", startTime.Unix()),
-			attribute.String("traceID", parentSpan.SpanContext().TraceID().String()),
-			attribute.String("authHeader", fields.authHeader),
-			attribute.String("secretToken", s.secretToken),
-			attribute.String("uKey", uKey),
-			attribute.Int64("slot", int64(fields.slot)),
-			attribute.String("blockHash", fields.blockHash),
-		)
-		span.End()
-	}()
-
-	if fields.client != nil {
-		clientURL = fields.client.SafeClient.URL
-	}
-
-	logMetric := NewLogMetric(
-		map[string]any{
-			"method":      preFetchPayload,
-			"receivedAt":  startTime,
-			"clientIP":    fields.clientIP,
-			"clientURL":   clientURL,
-			"reqID":       id,
-			"traceID":     parentSpan.SpanContext().TraceID().String(),
-			"secretToken": s.secretToken,
-			"authHeader":  fields.authHeader,
-			"uKey":        uKey,
-			"slot":        int64(fields.slot),
-			"blockHash":   fields.blockHash,
-		},
-	)
-
-	s.logger.Info().Fields(logMetric.GetFields()).Msg("received preFetchGetPayload")
-
-	// If necessary, fetch the Optimistic V3 payload directly from the specified builder URL(s)
-	if fields.payloadFetchUrl != "" {
-		s.prefetchPayloadFromBuilder(ctx, spanctx, &fields, logMetric.Copy())
-		return
-	}
-
-	s.prefetchPayloadGRPC(ctx, spanctx, &fields, logMetric.Copy(), span, id, startTime)
-}
-
-func (s *Service) prefetchPayloadGRPC(
-	ctx context.Context,
-	spanctx context.Context,
-	fields *preFetcherFields,
-	logMetric *LogMetric,
-	span trace.Span,
-	reqID string,
-	startTime time.Time,
-) {
-	req := &relaygrpc.PreFetchGetPayloadRequest{
-		ReqId:       reqID,
-		Version:     s.version,
-		SecretToken: s.secretToken,
-		Slot:        fields.slot,
-		ParentHash:  fields.parentHash,
-		BlockHash:   fields.blockHash,
-		Pubkey:      fields.proposerPubKey,
-		ClientIp:    fields.clientIP,
-		ReceivedAt:  timestamppb.New(startTime),
-	}
-
-	var (
-		errChan            = make(chan *ErrorResp, len(s.clients)+2)
-		respChan           = make(chan *relaygrpc.PreFetchGetPayloadResponse, len(s.clients)+2)
-		payloadCacheKey    = common.GetKeyForCachingPayload(fields.slot, fields.parentHash, fields.blockHash, fields.proposerPubKey)
-		wg                 sync.WaitGroup
-		prefetchedRequests = 0
-		succeeds           = false
-	)
-
-	// Goroutine to handle cache
-	prefetchedRequests += 1
-	if s.getPayloadResponseForProxySlot == nil {
-		s.logger.Error().Fields(logMetric.GetFields()).Msg("PreFetchGetPayload :: cache is nil")
-		errChan <- toErrorResp(http.StatusInternalServerError, "cache is nil")
-	}
-
-	if cachedValue, exists := s.getPayloadResponseForProxySlot.Get(payloadCacheKey); exists && cachedValue != nil {
-		payloadResponseForProxy, ok := cachedValue.(*common.PayloadResponseForProxy)
-		if !ok {
-			s.logger.Error().Fields(logMetric.GetFields()).Msg("failed to cast cached value to GetPayloadResponseForProxy")
-			errChan <- toErrorResp(http.StatusInternalServerError, "failed to cast cached value")
-		}
-		marshaledVal, err := payloadResponseForProxy.GetMarshalledResponse()
-		if err != nil {
-			s.logger.Error().Fields(logMetric.GetFields()).Msg("failed to marshal cached value to GetPayloadResponseForProxy")
-			errChan <- toErrorResp(http.StatusInternalServerError, "failed to marshal cached value")
-		}
-
-		resp := &relaygrpc.PreFetchGetPayloadResponse{
-			Code:                      uint32(codes.OK),
-			Message:                   "Pre fetch getPayload succeeded",
-			VersionedExecutionPayload: marshaledVal,
-		}
-
-		s.logger.Info().Fields(logMetric.GetFields()).Msg("PreFetchGetPayload-cache hit")
-
-		respChan <- resp
-		succeeds = true
-	} else {
-		errChan <- toErrorResp(http.StatusBadRequest, "local payload not found")
-	}
-
-	if !succeeds {
-		clients := s.clients
-		if fields.client != nil {
-			clients = append(clients, fields.client)
-		}
-
-		// Goroutines to fetch payloads
-		prefetchedRequests += len(clients)
-		for _, client := range clients {
-			wg.Add(1)
-			go func(client *common.ParentClient) {
-				defer wg.Done()
-				prefetchLogger := s.logger.With().Fields(logMetric.GetFields()).Logger()
-				s.prefetchPayload(ctx, spanctx, client.SafeClient, req, span, errChan, respChan, prefetchLogger)
-			}(client)
-		}
-	}
-	// Wait for all goroutines to finish
-	defer func() {
-		go func() {
-			wg.Wait()
-			close(respChan)
-			close(errChan)
-		}()
-	}()
-
-	// Process responses
-	for i := 0; i < prefetchedRequests; i++ {
-		select {
-		case <-ctx.Done():
-			s.logger.Warn().Fields(logMetric.GetFields()).Msg("PreFetchGetPayload :: Context canceled")
-		case _err := <-errChan:
-			s.logger.Error().Fields(logMetric.GetFields()).Interface("error", _err).Msg("PreFetchGetPayload :: Received error")
-		case out := <-respChan:
-			proxyCacheKey := common.GetKeyForCachingPayload(fields.slot, fields.parentHash, fields.blockHash, fields.proposerPubKey)
-
-			payloadResponse := &common.PayloadResponseForProxy{
-				MarshalledPayloadResponse: out.VersionedExecutionPayload,
-				BlockValue:                fields.blockValue,
-			}
-
-			if err := s.getPayloadResponseForProxySlot.Add(proxyCacheKey, payloadResponse, cache.DefaultExpiration); err != nil {
-				s.logger.Warn().Fields(logMetric.GetFields()).Err(err).Msg("PreFetchGetPayload :: respChan :: cache execution payload failed")
-				return
-			}
-
-			s.logger.Info().Fields(logMetric.GetFields()).Msg("PreFetchGetPayload :: respChan :: preFetchGetPayload succeeded")
-			return
-		}
-	}
-}
-
-func (s *Service) prefetchPayloadFromBuilder(ctx context.Context, spanCtx context.Context, fields *preFetcherFields, logMetric *LogMetric) {
-	_, span := s.tracer.Start(ctx, "prefetchPayloadFromBuilder")
-	var success atomic.Bool
-
-	defer func() {
-		span.SetAttributes(attribute.Bool("success", success.Load()))
-		span.End()
-	}()
-
-	payloadUrlsData := common.SafeSplit(fields.payloadFetchUrl, common.PayloadUrlsTypeSeparator)
-
-	if len(payloadUrlsData) != common.PayloadUrlsDataExpectedLength {
-		logMetric.Fields(map[string]any{"payloadUrlsData": payloadUrlsData})
-		s.logger.Error().Err(errors.New("invalid payload URL format")).Fields(logMetric.GetFields()).Msg("Failed to fetch Optimistic V3 payload from builder")
-		return
-	}
-
-	payloadUrlType := payloadUrlsData[common.PayloadUrlTypeIndex]
-	payloadUrlsCSV := payloadUrlsData[common.PayloadUrlsCSVIndex]
-	payloadUrls := common.SafeSplit(payloadUrlsCSV, ",")
-
-	logMetric.Fields(map[string]any{
-		"payloadUrlsData": payloadUrlsData,
-		"payloadUrls":     payloadUrls,
-	})
-
-	span.SetAttributes(
-		attribute.String("payloadUrlType", payloadUrlType),
-		attribute.StringSlice("payloadUrls", payloadUrls),
-	)
-
-	switch common.PayloadUrlType(payloadUrlType) {
-	case common.PayloadUrlTypeHTTP:
-		success.Store(s.clientPreFetchGetPayloadHTTP(ctx, logMetric, fields, payloadUrls))
-		return
-	case common.PayloadUrlTypeGRPC:
-		// We only support HTTP requests for Optimistic V3 payloads from builders for now
-		s.logger.Warn().Fields(logMetric.GetFields()).Msg("Ignoring fetch Optimistic V3 payload request with 'grpc' URL type")
-		return
-	default:
-		s.logger.Error().Err(errors.New("invalid payload URL type")).Fields(logMetric.GetFields()).Msg("Failed to fetch Optimistic V3 payload from builder")
-		return
-	}
-}
-
-func (s *Service) clientPreFetchGetPayloadHTTP(
-	ctx context.Context,
-	logMetric *LogMetric,
-	fields *preFetcherFields,
-	payloadUrls []string,
-) bool {
-	_, fetchSpan := s.tracer.Start(ctx, "clientPreFetchGetPayloadHTTP")
-	defer func() {
-		fetchSpan.SetAttributes(
-			attribute.Int64("slot", int64(fields.slot)),
-			attribute.String("blockHash", fields.blockHash),
-			attribute.String("parentHash", fields.parentHash),
-			attribute.String("proposerPubkey", fields.proposerPubKey),
-			attribute.String("builderPubkey", fields.builderPubKey),
-		)
-		fetchSpan.End()
-	}()
-
-	payload, err := s.prepareGetPayloadV3Request(fields.blockHash)
-	if err != nil {
-		s.logger.Error().Err(err).Fields(logMetric.GetFields()).Msg("failed to prepare HTTP get_payload_v3 request")
-		return false
-	}
-
-	responseChan := make(chan *common.VersionedSubmitBlockRequest, len(payloadUrls))
-
-	// Send request to all builders
-	for _, payloadUrl := range payloadUrls {
-		url := payloadUrl + common.PathGetPayloadV3
-
-		go func() {
-			result := new(common.VersionedSubmitBlockRequest)
-			code, durationMS, err := httpclient.FetchSSZ(http.MethodPost, url, payload, result, nil, true)
-
-			// TODO: should we try with JSON if ssz fails?
-			if err != nil {
-				s.logger.Error().
-					Fields(logMetric.GetFields()).
-					Err(err).Str("url", url).
-					Int("code", code).
-					Int64("durationMS", durationMS).
-					Msg("failed to prefetch payload with HTTP")
-				return
-			}
-
-			// Send to response channel
-			responseChan <- result
-		}()
-	}
-
-	// Process first positive response from builder (or timeout)
-	return s.processGetPayloadV3Responses(ctx, responseChan, logMetric, fields)
-}
-
-func (s *Service) prepareGetPayloadV3Request(blockHash string) (*common.SignedGetPayloadV3, error) {
-	getPayloadV3 := &common.GetPayloadV3{
-		BlockHash:      phase0.Hash32(gethcommon.HexToHash(blockHash)),
-		RequestTs:      uint64(time.Now().UnixMilli()),
-		RelayPublicKey: s.publicKey,
-	}
-
-	signature, err := ssz.SignMessage(getPayloadV3, s.builderSigningDomain, s.secretKey)
-	if err != nil {
-		return nil, err
-	}
-
-	return &common.SignedGetPayloadV3{
-		Message:   getPayloadV3,
-		Signature: signature,
-	}, nil
-}
-
-func (s *Service) processGetPayloadV3Responses(
-	ctx context.Context,
-	responseChan chan *common.VersionedSubmitBlockRequest,
-	logMetric *LogMetric,
-	fields *preFetcherFields,
-) bool {
-	for {
-		select {
-		case <-ctx.Done():
-			s.logger.Error().Fields(logMetric.GetFields()).Msg("PreFetchGetPayloadV3 :: context cancelled")
-			return false
-		case response := <-responseChan:
-			if response == nil {
-				s.logger.Error().Fields(logMetric.GetFields()).Msg("PreFetchGetPayloadV3 :: failed to prefetch payload with HTTP, received nil payload from builder")
-				continue
-			}
-
-			getPayloadResponseSpec, err := common.BuildGetPayloadResponse(response)
-			if err != nil {
-				s.logger.Fatal().Fields(logMetric.GetFields()).Err(err)
-			}
-
-			getPayloadResponse := common.VersionedSubmitBlindedBlockResponse{VersionedSubmitBlindedBlockResponse: *getPayloadResponseSpec}
-			proxyCacheKey := common.GetKeyForCachingPayload(fields.slot, fields.parentHash, fields.blockHash, fields.proposerPubKey)
-
-			payloadResponse := &common.PayloadResponseForProxy{
-				PayloadResponse: getPayloadResponse,
-				BlockValue:      fields.blockValue,
-			}
-
-			if err := s.getPayloadResponseForProxySlot.Add(proxyCacheKey, payloadResponse, cache.DefaultExpiration); err != nil {
-				s.logger.Warn().Fields(logMetric.GetFields()).Err(err).Msg("PreFetchGetPayloadV3 :: cache execution payload already exists")
-				return true
-			}
-
-			s.logger.Info().Fields(logMetric.GetFields()).Msg("PreFetchGetPayloadV3 :: preFetchGetPayload succeeded")
-			return true
-		case <-time.After(common.OptimisticV3FetchPayloadTimeout):
-			s.logger.Error().Fields(logMetric.GetFields()).Msg("PreFetchGetPayloadV3 :: timeout waiting for prefetch payload HTTP response")
-			return false
-		}
-	}
-}
-
-func (s *Service) prefetchPayloadToSignedBlindedBeaconBlock(ctx context.Context, payload []byte) (*common.VersionedSignedBlindedBeaconBlock, *ErrorResp) {
-	_, readPayload := s.tracer.Start(ctx, "validateAndFetchPayload-readPayload")
-
-	bodyString := string(payload)
-	blockHashIndex := strings.LastIndex(bodyString, "\"block_hash\"")
-
-	if blockHashIndex == -1 {
-		return nil, toErrorResp(http.StatusBadRequest, "invalid input")
-	}
-	readPayload.End(trace.WithTimestamp(time.Now()))
-
-	_, decodeJSONSpan := s.tracer.Start(ctx, "validateAndFetchPayload-decodeJSON")
-	signedBlindedBeaconBlock, err := fastjson.UnmarshalToSignedBlindedBeaconBlock(bodyString)
-	if err != nil {
-		decodeJSONSpan.End(trace.WithTimestamp(time.Now()))
-		return nil, toErrorResp(http.StatusBadRequest, err.Error())
-	}
-	decodeJSONSpan.End(trace.WithTimestamp(time.Now()))
-	return signedBlindedBeaconBlock, nil
-}
-
-func (s *Service) validateAndFetchPayload(ctx context.Context, signedBlindedBeaconBlock *common.VersionedSignedBlindedBeaconBlock) (*common.VersionedPayloadInfo, *ErrorResp) {
-	slot, err := signedBlindedBeaconBlock.Slot()
-	if err != nil {
-		return nil, toErrorResp(http.StatusBadRequest, "failed to get slot")
-	}
-
-	blockHash, err := signedBlindedBeaconBlock.ExecutionBlockHash()
-	if err != nil {
-		return nil, toErrorResp(http.StatusBadRequest, "failed to get block hash")
-	}
-	blockHashString := blockHash.String()
-
-	parentHash, err := signedBlindedBeaconBlock.ExecutionParentHash()
-	if err != nil {
-		return nil, toErrorResp(http.StatusBadRequest, "failed to get parent hash")
-	}
-
-	_, checkRequestTimingSpan := s.tracer.Start(ctx, "validateAndFetchPayload-checkRequestTiming")
-
-	slotStartTime := GetSlotStartTime(s.beaconGenesisTime, int64(slot), s.secondsPerSlot)
-	msIntoSlot := time.Since(slotStartTime).Milliseconds()
-
-	if msIntoSlot < 0 {
-		_msSinceSlotStart := time.Now().UTC().UnixMilli() - slotStartTime.UnixMilli()
-		if _msSinceSlotStart < 0 {
-			delayMillis := (_msSinceSlotStart * -1) + int64(rand.Intn(50))
-			time.Sleep(time.Duration(delayMillis) * time.Millisecond)
-		}
-	} else if msIntoSlot > int64(getPayloadRequestCutoffMs) {
-		checkRequestTimingSpan.End(trace.WithTimestamp(time.Now()))
-		return nil, toErrorResp(http.StatusBadRequest, "timestamp too late")
-	}
-	checkRequestTimingSpan.End(trace.WithTimestamp(time.Now()))
-
-	_, fetchProposerForSlotSpan := s.tracer.Start(ctx, "validateAndFetchPayload-fetchProposerForSlot")
-	miniSlotDuty, err := s.IDataService.GetSlotDuty(uint64(slot))
-	if err != nil || miniSlotDuty == nil {
-		return nil, toErrorResp(http.StatusBadRequest, fmt.Sprintf("slot %v not found in memory", slot))
-	}
-	pub := miniSlotDuty.Registration.Message.Pubkey
-	pubkeyStr := pub.String()
-
-	fetchProposerForSlotSpan.End(trace.WithTimestamp(time.Now()))
-
-	_, verifySignatureSpan := s.tracer.Start(ctx, "validateAndFetchPayload-verifySignature")
-	ok, err := fastjson.CheckProposerSignature(s.ethNetworkDetails, signedBlindedBeaconBlock, pub[:])
-	if !ok || err != nil {
-		verifySignatureSpan.End(trace.WithTimestamp(time.Now()))
-		return nil, toErrorResp(http.StatusBadRequest, "invalid signature")
-	}
-	verifySignatureSpan.End(trace.WithTimestamp(time.Now()))
-
-	_, fetchPayloadFromCacheSpan := s.tracer.Start(ctx, "validateAndFetchPayload-fetchPayloadFromCache")
-	proxyCacheKey := common.GetKeyForCachingPayload(uint64(slot), parentHash.String(), blockHashString, pubkeyStr)
-	defer fetchPayloadFromCacheSpan.End()
-
-	var payloadResponse *common.PayloadResponseForProxy
-	var found bool
-
-	for i := 0; i < 30; i++ { // try for 1.5s with 50ms interval
-		if val, ok := s.getPayloadResponseForProxySlot.Get(proxyCacheKey); ok {
-			casted, castOk := val.(*common.PayloadResponseForProxy)
-			if castOk {
-				payloadResponse = casted
-				found = true
-				break
-			}
-			break
-		}
-		// not found, wait and retry
-		time.Sleep(50 * time.Millisecond)
-	}
-	if found {
-		versionedPayloadInfo, err := payloadResponse.BuildVersionedPayloadInfo(uint64(slot), parentHash.String(), blockHashString, pubkeyStr)
-		if err != nil {
-			return nil, toErrorResp(http.StatusOK, "failed to build versioned payload info")
-		}
-		return versionedPayloadInfo, nil
-	}
-
-	return &common.VersionedPayloadInfo{
-		Slot:       uint64(slot),
-		ParentHash: parentHash.String(),
-		BlockHash:  blockHashString,
-		Pubkey:     pubkeyStr,
-	}, toErrorResp(http.StatusOK, "pre fetch payload not available in cache after retries")
-
-}
-
-func (s *Service) GetPayload(ctx context.Context, log *zerolog.Logger, in *PayloadRequestParams) (*common.VersionedPayloadInfo, error) {
-	startTime := time.Now().UTC()
-	id := uuid.NewString()
-
-	parentSpan := trace.SpanFromContext(ctx)
-	ctx = trace.ContextWithSpan(context.Background(), parentSpan)
-
-	authKey := s.authKey
-	if in.AuthHeader != "" {
-		authKey = in.AuthHeader
-	}
-	ctx = metadata.AppendToOutgoingContext(ctx, "authorization", authKey)
-
-	*log = log.With().
-		Str("method", getPayload).
-		Time("receivedAt", in.ReceivedAt).
-		Str("reqID", id).
-		Bool("isAuthHeaderProvided", in.AuthHeader != "").
-		Logger()
-	log.Info().Msg("received getPayloadTrusted")
-	ctx, span := s.tracer.Start(ctx, "getPayload-start")
-	var (
-		slotInt       int64
-		blockHashStr  string
-		parentHashStr string
-		blockValueStr string
-		uKey          string
-		latency       int64
-	)
-	defer func() {
-		_, logTimingSpan := s.tracer.Start(ctx, "getPayload-logTimingSpan")
-		parentSpan.SetAttributes(
-			attribute.String("method", getPayload),
-			attribute.String("reqID", id),
-			attribute.Int64("receivedAt", in.ReceivedAt.Unix()),
-			attribute.Int64("slot", slotInt),
-			attribute.String("blockHash", blockHashStr),
-			attribute.String("parentHash", parentHashStr),
-			attribute.String("blockValue", blockValueStr),
-			attribute.String("uniqueKey", uKey),
-			attribute.Int64("latency", latency),
-		)
-		log.Info().Msg("added spans getPayloadTrusted")
-		logTimingSpan.End()
-		span.End()
-	}()
-
-	_, timeToRelayRequestSpan := s.tracer.Start(ctx, "getPayload-TimeToRelayRequest")
-
-	if in.GetPayloadStartTimeUnixMS != "" {
-		if getPayloadStartTime, err := strconv.ParseInt(in.GetPayloadStartTimeUnixMS, 10, 64); err == nil {
-			latency = in.ReceivedAt.Sub(time.UnixMilli(getPayloadStartTime)).Milliseconds()
-		} else {
-			log.Warn().Err(err).Msg("failed to parse getPayloadStartTimeUnixMS")
-		}
-	}
-
-	req := &relaygrpc.GetPayloadRequest{
-		ReqId:       id,
-		Payload:     in.Payload,
-		ClientIp:    in.ClientIP,
-		Version:     s.version,
-		ReceivedAt:  timestamppb.New(in.ReceivedAt),
-		SecretToken: s.secretToken,
-	}
-	timeToRelayRequestSpan.End()
-
-	_, prefetchPayloadToSignedBlindedBeaconBlockSpan := s.tracer.Start(ctx, "getPayload-prefetchPayloadToSignedBlindedBeaconBlockSpan")
-	blindedBeaconBlock, errRes := s.prefetchPayloadToSignedBlindedBeaconBlock(ctx, in.Payload)
-	if errRes != nil {
-		log.Error().Err(errRes).Msg("prefetchPayloadToSignedBlindedBeaconBlock failed")
-		go s.sendPayloadStats(in.Payload, log, false, nil, in.ReceivedAt, startTime, time.Now(), 0, id, in.ClientIP, in.ValidatorID, in.AccountID, latency, in.Cluster, in.UserAgent, in.SlotUID)
-		return nil, errRes
-	}
-	slot, err := blindedBeaconBlock.Slot()
-	if err != nil {
-		return nil, toErrorResp(http.StatusBadRequest, "failed to get slot")
-	}
-	blockHash, err := blindedBeaconBlock.ExecutionBlockHash()
-	if err != nil {
-		return nil, toErrorResp(http.StatusBadRequest, "failed to get block hash")
-	}
-	parentHash, err := blindedBeaconBlock.ExecutionParentHash()
-	if err != nil {
-		return nil, toErrorResp(http.StatusBadRequest, "failed to get parent hash")
-	}
-	slotInt = int64(slot)
-	blockHashStr = blockHash.String()
-	parentHashStr = parentHash.String()
-	uKey = fmt.Sprintf("slot_%v_bHash_%v_pHash_%v", slotInt, blockHashStr, parentHashStr)
-	*log = log.With().
-		Int64("Slot", slotInt).
-		Str("blockHash", blockHashStr).
-		Str("parentHash", parentHashStr).
-		Str("uKey", uKey).
-		Logger()
-	prefetchPayloadToSignedBlindedBeaconBlockSpan.End()
-
-	go func() {
-		if s.OnPayloadDelivered == nil {
-			log.Warn().Msg("skipping OnPayloadDelivered")
-			return
-		}
-		var pubkeyStr string
-		miniSlotDuty, err := s.IDataService.GetSlotDuty(uint64(slot))
-		if err == nil && miniSlotDuty != nil && miniSlotDuty.Registration != nil && miniSlotDuty.Registration.Message != nil {
-			pub := miniSlotDuty.Registration.Message.Pubkey
-			pubkeyStr = pub.String()
-		}
-		proposerRequestStartTimeUnixMS, _ := strconv.ParseInt(in.GetPayloadStartTimeUnixMS, 10, 64)
-		err = s.OnPayloadDelivered(uint64(slot), blockHashStr, parentHashStr, pubkeyStr, in.ClientIP, in.ReceivedAt, &blindedBeaconBlock.VersionedSignedBlindedBeaconBlock, proposerRequestStartTimeUnixMS, in.ValidatorID)
-		if err != nil {
-			log.Error().Err(err).Msg("failed to call OnPayloadDelivered")
-		}
-	}()
-
-	payloadInfoChan := make(chan *common.VersionedPayloadInfo, 1)
-
-	// validate and  fetch payload from cache
-	go func(ctx context.Context, l zerolog.Logger, parent trace.Span) {
-		ctx, childSpan := s.tracer.Start(ctx, "validateAndFetchPayload")
-		defer childSpan.End()
-
-		payloadInfo, err := s.validateAndFetchPayload(ctx, blindedBeaconBlock)
-		if err == nil && payloadInfo != nil {
-			select {
-			case payloadInfoChan <- payloadInfo:
-			default:
-			}
-			if err != nil {
-				l.Warn().Err(err).Msg("validateAndFetchPayload returned payload with partial error")
-			}
-		} else {
-			l.Warn().Err(err).Msg("validateAndFetchPayload returned no payload")
-		}
-	}(ctx, *log, parentSpan)
-
-	// fetch payload relay
-	for _, client := range s.clients {
-		go func(c *common.ParentClient, parent trace.Span) {
-			ctx, childSpan := s.tracer.Start(ctx, "getPayloadWithRetry")
-			defer childSpan.End()
-
-			resp, err := s.getPayloadWithRetry(ctx, c.SafeClient, childSpan, req, maxGetPayloadRetry)
-			if err == nil && resp != nil {
-				select {
-				case payloadInfoChan <- resp:
-				default:
-				}
-			}
-		}(client, parentSpan)
-	}
-
-	select {
-	case payloadInfo := <-payloadInfoChan:
-		go func() {
-			if s.blockPublishFunc != nil {
-				s.blockPublishFunc(s.tracer, s.logger, payloadInfo, blindedBeaconBlock, s.blockPublishingGatewayClient, s.gatewayAuthKey)
-			}
-		}()
-		slotStartTime := GetSlotStartTime(s.beaconGenesisTime, int64(payloadInfo.Slot), s.secondsPerSlot)
-		msIntoSlot := in.ReceivedAt.Sub(slotStartTime).Milliseconds()
-		duration := time.Since(startTime)
-		go s.sendPayloadStats(in.Payload, log, true, payloadInfo, in.ReceivedAt, startTime, slotStartTime, msIntoSlot, id, in.ClientIP, in.ValidatorID, in.AccountID, latency, in.Cluster, in.UserAgent, in.SlotUID)
-		blockValueStr = payloadInfo.GetBlockValue()
-		*log = log.With().
-			Int64("latency", latency).
-			Dur("duration", duration).
-			Int64("slotStartTime", slotStartTime.UnixMilli()).
-			Int64("msIntoSlot", msIntoSlot).
-			Str("blockValue", blockValueStr).
-			Logger()
-
-		return payloadInfo, nil
-	case <-time.After(1500 * time.Millisecond):
-	}
-	log.Error().Msg("timeout waiting for payload response")
-	go s.sendPayloadStats(in.Payload, log, false, nil, in.ReceivedAt, startTime, time.Now(), 0, id, in.ClientIP, in.ValidatorID, in.AccountID, latency, in.Cluster, in.UserAgent, in.SlotUID)
-	return nil, toErrorResp(http.StatusOK, "pre fetch payload not available in cache after retries")
-}
-
-type ErrorRespWithPayload struct {
-	err  *ErrorResp
-	resp *common.VersionedPayloadInfo
-}
-
-func (s *Service) getPayloadWithRetry(ctx context.Context, c *common.Client, parentSpan trace.Span, req *relaygrpc.GetPayloadRequest, retryCount int) (*common.VersionedPayloadInfo, *ErrorResp) {
-	for attempt := 0; attempt <= retryCount; attempt++ {
-		_, clientGetPayloadSpan := s.tracer.Start(ctx, "getPayloadWithRetry-getPayloadForClient")
-		clientCtx, cancel := context.WithTimeout(ctx, 3*time.Second)
-		resp, err := c.GetPayload(clientCtx, req)
-		cancel()
-		clientGetPayloadSpan.End(trace.WithTimestamp(time.Now()))
-
-		if err == nil && resp != nil {
-			if resp.GetCode() == uint32(codes.OK) {
-				return common.BuildVersionedPayloadInfoFromGrpcResponse(resp), nil
-			}
-			if resp.GetMessage() != "could not find requested payload" {
-				parentSpan.SetStatus(otelcodes.Error, resp.Message)
-				uKey := fmt.Sprintf("slot_%v_bHash_%v_pHash_%v", resp.GetSlot(), resp.GetBlockHash(), resp.GetParentHash())
-				clientGetPayloadSpan.SetAttributes(
-					attribute.String("relayError", resp.Message),
-					attribute.String("url", c.URL),
-					attribute.Int64("slot", int64(resp.GetSlot())),
-					attribute.String("BlockHash", resp.GetBlockHash()),
-					attribute.String("in.ParentHash", resp.GetParentHash()),
-					attribute.String("BlockValue", resp.GetBlockValue()),
-					attribute.String("uniqueKey", uKey),
-				)
-				return common.BuildVersionedPayloadInfoFromGrpcResponse(resp), toErrorResp(http.StatusBadRequest, resp.Message)
-			}
-		}
-
-		if attempt < retryCount {
-			time.Sleep(getPayloadInterval)
-			continue
-		}
-
-		if err != nil {
-			parentSpan.SetStatus(otelcodes.Error, err.Error())
-			return nil, toErrorResp(http.StatusInternalServerError, err.Error())
-		}
-		if resp == nil {
-			parentSpan.SetStatus(otelcodes.Error, "empty response from relay")
-			return nil, toErrorResp(http.StatusInternalServerError, "empty response from relay")
-		}
-
-		parentSpan.SetStatus(otelcodes.Error, resp.Message)
-		uKey := fmt.Sprintf("slot_%v_bHash_%v_pHash_%v", resp.GetSlot(), resp.GetBlockHash(), resp.GetParentHash())
-		clientGetPayloadSpan.SetAttributes(
-			attribute.String("relayError", resp.Message),
-			attribute.String("url", c.URL),
-			attribute.Int64("slot", int64(resp.GetSlot())),
-			attribute.String("BlockHash", resp.GetBlockHash()),
-			attribute.String("in.ParentHash", resp.GetParentHash()),
-			attribute.String("BlockValue", resp.GetBlockValue()),
-			attribute.String("uniqueKey", uKey),
-		)
-		return common.BuildVersionedPayloadInfoFromGrpcResponse(resp), toErrorResp(http.StatusBadRequest, resp.Message)
-	}
-
-	return nil, toErrorResp(http.StatusInternalServerError, "all relay retries failed")
-}
-
-func (s *Service) sendPayloadStats(payload []byte, log *zerolog.Logger, isSucceeded bool, resp *common.VersionedPayloadInfo, receivedAt, startTime, slotStartTime time.Time, msIntoSlot int64, id, clientIP, validatorID, accountID string, latency int64, cluster, userAgent, slotUID string) {
-	// 3 different scenario calling sendPayload stats
-	// case 1 : resp success
-	// case 2: Err case with resp
-	// case 2: Err case with no resp
-	out := resp.Copy()
-	if out.GetSlot() != 0 {
-		slotStartTime = GetSlotStartTime(s.beaconGenesisTime, int64(out.GetSlot()), s.secondsPerSlot)
-		msIntoSlot = receivedAt.Sub(slotStartTime).Milliseconds()
-	}
-	if out == nil {
-		decodedPayload := new(common.VersionedSignedBlindedBeaconBlock)
-		if err := json.NewDecoder(bytes.NewReader(payload)).Decode(decodedPayload); err != nil {
-			log.Warn().Msg("failed to decode getPayload request")
-			return
-		} else {
-			_slot, err := decodedPayload.Slot()
-			if err != nil {
-				log.Warn().Err(err).Msg("failed to decode getPayload slot")
-				return
-			} else {
-				out = new(common.VersionedPayloadInfo)
-				out.SetSlot(uint64(_slot))
-				slotStartTime = GetSlotStartTime(s.beaconGenesisTime, int64(out.Slot), s.secondsPerSlot)
-				msIntoSlot = receivedAt.Sub(slotStartTime).Milliseconds()
-				_blockHash, err := decodedPayload.ExecutionBlockHash()
-				if err != nil {
-					log.Warn().Err(err).Msg("failed to decode getPayload BlockHash")
-				} else {
-					out.SetBlockHash(_blockHash.String())
-					parentHash, err := decodedPayload.ExecutionParentHash()
-					if err != nil {
-						log.Warn().Err(err).Msg("failed to decode getPayload parentHash")
-					} else {
-						out.SetParentHash(parentHash.String())
-					}
-				}
-			}
-		}
-	}
-
-	statsUserAgent := userAgent
-	if cluster != "" {
-		statsUserAgent = fmt.Sprintf("%s/%s", statsUserAgent, cluster)
-	}
-
-	statsRecord := SlotStatsRecord{
-		PayloadReqID:              id,
-		PayloadReqReceivedAt:      receivedAt,
-		PayloadReqDuration:        time.Since(startTime),
-		PayloadReqDurationInMs:    time.Since(startTime).Milliseconds(),
-		PayloadMsIntoSlot:         msIntoSlot,
-		PayloadSucceeded:          isSucceeded,
-		PayloadDeliveredBlockHash: out.GetBlockHash(),
-		PayloadBlockValue:         out.GetBlockValue(),
-		PayloadUserAgent:          statsUserAgent,
-		Slot:                      out.GetSlot(),
-		ParentHash:                out.GetParentHash(),
-		PubKey:                    out.GetPubkey(),
-		SlotStartTime:             slotStartTime,
-		ClientIP:                  clientIP,
-		NodeID:                    s.nodeID,
-		AccountID:                 accountID,
-		ValidatorID:               validatorID,
-		GetPayloadLatency:         latency,
-		PayloadSlotUID:            slotUID,
-	}
-	var (
-		isRelayProxyWin bool
-		//isSlotUIDMatch  bool
-		fallback SlotStatsRecord
-	)
-	k := fmt.Sprintf("slot-%v-parentHash-%v", out.GetSlot(), out.GetParentHash())
-	v, ok := s.slotStats.Get(k)
-	if ok {
-		if records, success := v.([]SlotStatsRecord); success {
-			for i, record := range records {
-				if i == len(records)-1 {
-					fallback = record
-				}
-				if record.HeaderDeliveredBlockHash == out.GetBlockHash() {
-					mergeSlotStats(&record, &statsRecord)
-					isRelayProxyWin = true
-					//isSlotUIDMatch = record.HeaderSlotUID == statsRecord.PayloadSlotUID
-					break
-				}
-				if !isRelayProxyWin {
-					mergeSlotStats(&fallback, &statsRecord)
-				}
-			}
-		}
-	} else {
-		log.Warn().Str("slotKey", k).Msg("no previous slot stats found, creating new record")
-	}
-	s.slotStatsEvent.Set(k, statsRecord, cache.DefaultExpiration) // replace with updated slot stats
-
-	if isRelayProxyWin {
-		log.Info().Str("slotKey", k).Msg("emit slot won event")
-		s.fluentD.LogToFluentD(fluentstats.Record{
-			Type: TypeRelayProxySlotWon,
-			Data: statsRecord,
-		}, time.Now().UTC(), s.nodeID, StatsRelayProxySlotWon)
-	}
-
-	payloadStats := GetPayloadStatsRecord{
-		RequestReceivedAt: receivedAt,
-		Duration:          time.Since(startTime),
-		SlotStartTime:     slotStartTime,
-		MsIntoSlot:        msIntoSlot,
-		Slot:              out.GetSlot(),
-		ParentHash:        out.GetParentHash(),
-		PubKey:            out.GetPubkey(),
-		BlockHash:         out.GetBlockHash(),
-		BlockValue:        out.GetBlockValue(),
-		ReqID:             id,
-		ClientIP:          clientIP,
-		Succeeded:         true,
-		NodeID:            s.nodeID,
-		AccountID:         accountID,
-		ValidatorID:       validatorID,
-		Latency:           latency,
-		UserAgent:         statsUserAgent,
-	}
-
-	s.fluentD.LogToFluentD(fluentstats.Record{
-		Type: TypeRelayProxyGetPayload,
-		Data: payloadStats,
-	}, time.Now().UTC(), s.nodeID, StatsRelayProxyGetPayload)
-}
-func mergeSlotStats(record *SlotStatsRecord, statsRecord *SlotStatsRecord) {
-	statsRecord.HeaderReqID = record.HeaderReqID
-	statsRecord.HeaderReqReceivedAt = record.HeaderReqReceivedAt
-	statsRecord.HeaderReqDuration = record.HeaderReqDuration
-	statsRecord.HeaderReqDurationInMs = record.HeaderReqDurationInMs
-	statsRecord.HeaderMsIntoSlot = record.HeaderMsIntoSlot
-	statsRecord.HeaderMsIntoSlotWithDelay = record.HeaderMsIntoSlotWithDelay
-	statsRecord.HeaderDelayInMs = record.HeaderDelayInMs
-	statsRecord.HeaderMaxDelayInMs = record.HeaderMaxDelayInMs
-	statsRecord.HeaderSucceeded = record.HeaderSucceeded
-	statsRecord.HeaderDeliveredBlockHash = record.HeaderDeliveredBlockHash
-	statsRecord.HeaderBlockValue = record.HeaderBlockValue
-	statsRecord.HeaderUserAgent = record.HeaderUserAgent
-	statsRecord.PubKey = record.PubKey
-	statsRecord.GetHeaderLatency = record.GetHeaderLatency
-	statsRecord.HeaderStartTimeUnixMs = record.HeaderStartTimeUnixMs
-	statsRecord.HeaderSlotUID = record.HeaderSlotUID
-
-	statsRecord.AccountID = record.AccountID
-	statsRecord.ValidatorID = record.ValidatorID
 }
 
 func (s *Service) keyForCachingBids(slot uint64, parentHash string, proposerPubkey string) string {
@@ -2670,214 +1405,6 @@ func (s *Service) logRecord(record SlotStatsRecord, slotKey string, userAgent st
 	}, time.Now().UTC(), s.nodeID, StatsRelayProxySlotStats)
 }
 
-func (s *Service) prefetchPayload(
-	ctx context.Context,
-	spanctx context.Context,
-	client *common.Client,
-	req *relaygrpc.PreFetchGetPayloadRequest,
-	span trace.Span,
-	errChan chan *ErrorResp,
-	respChan chan *relaygrpc.PreFetchGetPayloadResponse,
-	logger zerolog.Logger,
-) {
-	clientCtx, cancel := context.WithTimeout(ctx, 3*time.Second)
-	defer cancel()
-	exitSignal := false
-	wg := &sync.WaitGroup{}
-	mu := &sync.Mutex{}
-	wg.Add(1)
-	go func() {
-		defer wg.Done()
-		for i := 0; i < 5 && !exitSignal; i++ {
-			childCtx, childSpan := s.tracer.Start(spanctx, "PreFetchGetPayload")
-			_, reqSpan := s.tracer.Start(childCtx, "PreFetchGetPayload-request")
-			childSpan.SetAttributes(
-				attribute.String("url", client.URL),
-				attribute.String("nodeID", client.NodeID),
-			)
-			out, err := client.PreFetchGetPayload(clientCtx, req)
-			reqSpan.End()
-			if exitSignal {
-				childSpan.End()
-				return
-			}
-			if err != nil {
-				logger.Error().
-					Err(err).
-					Str("url", client.URL).
-					Msg("prefetchPayload: error fetching payload")
-				span.SetStatus(otelcodes.Error, err.Error())
-				time.Sleep(100 * time.Millisecond)
-				childSpan.End()
-				continue
-			}
-
-			if out == nil {
-				logger.Error().
-					Str("url", client.URL).
-					Msg("prefetchPayload: received nil payload from relay")
-				span.SetStatus(otelcodes.Error, "nil payload")
-				time.Sleep(100 * time.Millisecond)
-				childSpan.End()
-				continue
-			}
-
-			if out.Code != uint32(codes.OK) {
-				logger.Error().
-					Uint32("code", out.Code).
-					Str("message", out.Message).
-					Str("url", client.URL).
-					Msg("prefetchPayload: invalid payload or failure response code")
-				span.SetStatus(otelcodes.Error, out.Message)
-				time.Sleep(100 * time.Millisecond)
-				childSpan.End()
-				continue
-			}
-
-			logger.Info().
-				Str("url", client.URL).
-				Msg("prefetchPayload: preFetchGetPayload succeeded")
-			mu.Lock()
-			if !exitSignal {
-				exitSignal = true
-				respChan <- out
-				cancel()
-			}
-			mu.Unlock()
-			childSpan.End()
-			return
-		}
-	}()
-
-	wg.Add(1)
-	go func() {
-		defer wg.Done()
-		for i := 0; i < 5 && !exitSignal; i++ {
-			reqCtx, childSpan := s.tracer.Start(spanctx, "PreFetchGetPayloadPlaceHTTPRequest")
-			childSpan.SetAttributes(
-				attribute.String("url", client.URL),
-				attribute.String("nodeID", client.NodeID),
-			)
-			out, err := s.PreFetchGetPayloadPlaceHTTPRequest(clientCtx, reqCtx, req, client.URL, client.NodeID)
-			if exitSignal {
-				childSpan.End()
-				return
-			}
-			if err != nil {
-				logger.Error().
-					Err(err).
-					Str("url", client.URL).
-					Msg("prefetchPayload: error fetching payload")
-				span.SetStatus(otelcodes.Error, err.Error())
-				time.Sleep(100 * time.Millisecond)
-				childSpan.End()
-				continue
-			}
-
-			if out == nil {
-				logger.Error().
-					Str("url", client.URL).
-					Msg("prefetchPayload: received nil payload from relay")
-				span.SetStatus(otelcodes.Error, "nil payload")
-				time.Sleep(100 * time.Millisecond)
-				childSpan.End()
-				continue
-			}
-
-			if out.Code != uint32(codes.OK) {
-				logger.Error().
-					Uint32("code", out.Code).
-					Str("message", out.Message).
-					Str("url", client.URL).
-					Msg("prefetchPayload: invalid payload or failure response code")
-				span.SetStatus(otelcodes.Error, out.Message)
-				time.Sleep(100 * time.Millisecond)
-				childSpan.End()
-				continue
-			}
-
-			logger.Info().
-				Str("url", client.URL).
-				Msg("prefetchPayload: preFetchGetPayload succeeded")
-			mu.Lock()
-			if !exitSignal {
-				exitSignal = true
-				respChan <- out
-				cancel()
-			}
-			mu.Unlock()
-			childSpan.End()
-			return
-		}
-	}()
-
-	wg.Wait()
-	if exitSignal {
-		return
-	}
-
-	errChan <- toErrorResp(http.StatusInternalServerError, "relay failed all attempts")
-}
-
-func (s *Service) PreFetchGetPayloadPlaceHTTPRequest(ctx context.Context, reqCtx context.Context, origReq *relaygrpc.PreFetchGetPayloadRequest, url string, nodeID string) (*relaygrpc.PreFetchGetPayloadResponse, error) {
-	reqData := common.PreFetchGetPayloadRequestHTTP{
-		Slot:       origReq.GetSlot(),
-		ParentHash: origReq.GetParentHash(),
-		BlockHash:  origReq.GetBlockHash(),
-		Pubkey:     origReq.GetPubkey(),
-		ClientIp:   origReq.GetClientIp(),
-		ReceivedAt: origReq.GetReceivedAt(),
-	}
-	_, marshalSpan := s.tracer.Start(reqCtx, "PreFetchGetPayloadPlaceHTTPRequest-marshal")
-	reqJSON, err := json.Marshal(reqData)
-	marshalSpan.End()
-	if err != nil {
-		return nil, err
-	}
-	originalURL := url
-	port := ":18555"
-
-	if strings.Contains(url, ":") {
-		host, portNumber, err := net.SplitHostPort(url)
-		if err != nil {
-			return nil, err
-		}
-		url = host
-		if portNumber == "5015" {
-			port = ":18550"
-		}
-	}
-
-	finalURL := "http://" + url + port + common.PathPrefetchBlock
-	s.logger.Info().Str("nodeID", nodeID).Str("finalURL", finalURL).Str("originalURL", originalURL).Msg("making prefetch request")
-	req, err := http.NewRequest("GET", finalURL, bytes.NewReader(reqJSON))
-	if err != nil {
-		return nil, err
-	}
-	req.Header.Set("Content-Type", "application/json")
-
-	client := &http.Client{}
-	_, requestSpan := s.tracer.Start(reqCtx, "PreFetchGetPayloadPlaceHTTPRequest-request")
-	resp, err := client.Do(req)
-	requestSpan.End()
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-
-	_, unmarshalSpan := s.tracer.Start(reqCtx, "PreFetchGetPayloadPlaceHTTPRequest-unmarshal")
-	var respData common.PreFetchGetPayloadResponseHTTP
-	if err := json.NewDecoder(resp.Body).Decode(&respData); err != nil && err != io.EOF {
-		unmarshalSpan.End()
-		return nil, err
-	}
-	unmarshalSpan.End()
-	return &relaygrpc.PreFetchGetPayloadResponse{
-		Code:                      respData.Code,
-		Message:                   respData.Message,
-		VersionedExecutionPayload: respData.VersionedExecutionPayload,
-	}, nil
-}
 func (s *Service) StartStreamSlotInfo(ctx context.Context, wg *sync.WaitGroup) {
 	for _, client := range s.streamingBlockClients {
 		wg.Add(1)
